resume-analyzer-free/requirements.txt
streamlit==1.39.0
python-docx==1.1.2
PyPDF2==3.0.1
python-dotenv==1.0.1
